Tag function migrated to common
