+++ lambda/common/s3.js
@@ -38,6 +38,26 @@ const getS3Keys = keyParams =>
     });
   });
 
+const getS3ObjectTags = s3Object =>
+  new Promise((resolve, reject) => {
+    const taggingParams = {
+      Bucket: trackBucket,
+      Key: s3Object.key
+    };
+    s3.getObjectTagging(taggingParams, function(err, trackTags) {
+      if (err) {
+        reject(err);
+      }
+      else {
+        trackTags.TagSet.forEach(trackTag => {
+          s3Object[`${trackTag.Key}`] = trackTag.Value;
+        });
+        resolve(s3Object);
+      }
+    });
+  });
+
 module.exports = {
-  getS3Keys
+  getS3Keys,
+  getS3ObjectTags
 };
